ar2en_service/requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
transformers==4.34.0
torch==2.0.1
pydantic==1.10.9
sentencepiece==0.1.96
sacremoses
kafka-python==2.0.2
numpy<2
